perf(data): lazy fetch plugin versions
Problem: vim.pack.get() performs IO on the lock file,
slowing down how long it takes to fetch and display
the plugins in the UI

Solution: lazy fetch version the first time `info.version`
is indexed and cache it in the `info` table for future accesses

Special case to this is `zpack` as it doesn't store info about
itself anywhere, so we have to fetch it from vim.pack.
This is still about ~100ms faster than fetching all plugins from
vim.pack, though.